вынес из UserData UserDataRow и OhoneBox
